Fix for no response when send USSD during airplane mode.
When Airplane mode is on and call is placed, create
connection will fail. "To place a call, first
turn off Airplane mode" will be shown by InCallUI.
However, when it's a USSD number, the call won't
be added to InCallService. InCallUI won't track
this kind of call. So when creating connection
fails, InCallUI won't be notified, and won't pop
the error dialog up.

Change-Id: I8be7e709fd88f3bc9262077262ffd27f4f1a7a75